refactor: bringing out of validator classes and helper network functions
