+++ util/validator.py
@@ -5,6 +5,8 @@ import socket
 from prompt_toolkit.document import Document
 from questionary import ValidationError, Validator
 
+from config.GlobalList import Blacklist, Whitelist
+
 ipv4 = re.compile(r"((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)(\.|$)){4}")
 
 
@@ -30,3 +32,41 @@ class IPValidator(Validator):
                 socket.inet_aton(text)
                 return text
         raise error
+
+
+class NameInBlacklist(Validator):
+    def validate(self, document: Document):
+        blacklist = Blacklist()
+        name = document.text
+        if blacklist.has(name):
+            raise ValidationError(
+                message="Name already in list", cursor_position=len(name)
+            )
+
+
+class NameInWhitelist(Validator):
+    def validate(self, document: Document):
+        whitelist = Whitelist()
+        name = document.text
+        if whitelist.has(name):
+            raise ValidationError(
+                message="Name already in list", cursor_position=len(name)
+            )
+
+
+class IPInBlacklist(Validator):
+    def validate(self, document: Document):
+        super().validate(document)
+        blacklist = Blacklist()
+        ip = document.text
+        if ip in blacklist:
+            raise ValidationError(message="IP already in list", cursor_position=len(ip))
+
+
+class IPInWhitelist(IPValidator):
+    def validate(self, document: Document):
+        super().validate(document)
+        whitelist = Whitelist()
+        ip = document.text
+        if ip in whitelist:
+            raise ValidationError(message="IP already in list", cursor_position=len(ip))
